[TM-878] Move the WorkdaysTable out of the DataTable directory.
