+++ src/components/extensive/WizardForm/types.ts
@@ -13,7 +13,6 @@ import { RHFLeadershipTeamTableProps } from "@/components/elements/Inputs/DataTa
 import { RHFOwnershipStakeTableProps } from "@/components/elements/Inputs/DataTable/RHFOwnershipStakeTable";
 import { RHFSeedingProps } from "@/components/elements/Inputs/DataTable/RHFSeedingTable";
 import { RHFStrataTableProps } from "@/components/elements/Inputs/DataTable/RHFStrataTable";
-import { RHFWorkdaysTableProps } from "@/components/elements/Inputs/DataTable/RHFWorkdaysTable";
 import { RHFDropdownProps } from "@/components/elements/Inputs/Dropdown/RHFDropdown";
 import { RHFFileInputProps } from "@/components/elements/Inputs/FileInput/RHFFileInput";
 import { InputProps } from "@/components/elements/Inputs/Input/Input";
@@ -24,6 +23,7 @@ import { RHFSelectImageProps } from "@/components/elements/Inputs/SelectImage/RH
 import { TextAreaProps } from "@/components/elements/Inputs/textArea/TextArea";
 import { RHFSeedingTableInputProps } from "@/components/elements/Inputs/TreeSpeciesInput/RHFSeedingTableInput";
 import { RHFTreeSpeciesInputProps } from "@/components/elements/Inputs/TreeSpeciesInput/RHFTreeSpeciesInput";
+import { RHFWorkdaysTableProps } from "@/components/elements/Inputs/WorkdaysInput/RHFWorkdaysTable";
 
 export interface FormStepSchema {
   title: string;
